Fleshing out error conditions, config, and callbacks
This is to support a bunch of tests in the next commit.  These changes
clean up a bunch of stuff that made the code insanely hard to test.

There's a little more work in server.coffee but I can now write some
pretty good tests for the event based log stuff and the server monitor