Improves server status feature
WARNING: There's a socket file descriptor leak in the Starbound server, verified
in OS X and Linux.  Every connection to the server (player or status check)
opens a socket but leaves a file open even when the connection is closed.
Eventually, the server will hit the limit of open files set by the OS for the
process.  At this point, the server will crash.

* Server status is now shown in Server Info panel
* Server Info panel moved to top of left column
* Enable/disable feature on configuration
* Configurable interval for checking