Moves some state/events into main server class
Trying to consolidate server state and events into one class.
The hope is that StarboundServer will track state from multiple "backends" such
as the log file, testing the game port, and perhaps by acting as a client in the
future.  This one entrypoint will provide the latest known state without
requiring knowledge of where that state came from.

The biggest problem being solved right now is that a server stop message in the
log file and the port not responding both emit stop events which could lead to
multiple messages in the chat log on the site.  By consolidating this,
StarboundServer knows that the server is already stopped when receiving the
second event and it will not emit it since we already know it's down!